* net_main.c: Fix NET_SendToAll() so that it really skips non-connected clients. While we're at it, change the type of the blocktime argument to double (the comparison is against a double.) Rename the cryptic state1 and state2 variables to something more indicative of thei purpose, ie. msg_init and msg_sent. from uhexen2. * net.h, host.c, sv_main.c: Adjust for the NET_SendToAll() argument type change.
git-svn-id: svn://svn.code.sf.net/p/quakespasm/code/trunk/quakespasm@200 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/net.h
@@ -293,7 +293,7 @@ int			NET_SendUnreliableMessage (struct qsocket_s *sock, sizebuf_t *data);
 // returns 1 if the message was sent properly
 // returns -1 if the connection died
 
-int			NET_SendToAll(sizebuf_t *data, int blocktime);
+int			NET_SendToAll(sizebuf_t *data, double blocktime);
 // This is a reliable *blocking* send to all attached clients.
